fix: memory leaks, FULL JOIN, SortMerge outer join, deadlock victim, NULL support
Memory & Ownership:
- Add NodeHandleGuard RAII class for IxNodeHandle lifecycle management
- Fix IxIndexHandle constructor: remove leaked buf allocation, add destructor
- Rename release_node_handle to decrement_page_count

Executor Fixes:
- IndexScanExecutor: add eval_cond/check_all_conds for non-index condition filtering
- AggregationExecutor: fix non-GROUP-BY path to process all aggregates, not just first
- NestedLoopJoinExecutor: implement FULL JOIN right-side unmatched row tracking
- SortMergeJoinExecutor: add LEFT JOIN support with NULL-padded unmatched rows
- Portal: route LEFT_JOIN to SortMergeJoin when equi-conditions exist

Transaction:
- find_deadlock_victim: select youngest transaction (highest txn_id) as victim
- Fix get_transaction to return nullptr instead of asserting for missing txns

Infrastructure:
- Add NULL literal support (Value::is_null_, set_null, executor type check skip)
- Implement ReconstructTuple and IsWriteWriteConflict MVCC stubs
- Fix open_db directory management (remove premature chdir(".."))
- CREATE INDEX: populate with existing table data
- Fix HAVING column lookup for table-qualified names

Warnings:
- Fix all sign-compare warnings in executors and sm_meta.h
- Fix log_manager.h format specifier and initialization order

+++ rmdb/src/execution/executor_delete.h
@@ -48,7 +48,7 @@ class DeleteExecutor : public AbstractExecutor {
                               .get();
                 char *key = new char[index.col_tot_len];
                 int offset = 0;
-                for (size_t j = 0; j < index.col_num; j++) {
+                for (size_t j = 0; j < static_cast<size_t>(index.col_num); j++) {
                     memcpy(key + offset, rec->data + index.cols[j].offset, index.cols[j].len);
                     offset += index.cols[j].len;
                 }
