+++ rmdb/src/execution/executor_update.h
@@ -50,7 +50,7 @@ class UpdateExecutor : public AbstractExecutor {
                               .get();
                 char *old_key = new char[index.col_tot_len];
                 int offset = 0;
-                for (size_t j = 0; j < index.col_num; j++) {
+                for (size_t j = 0; j < static_cast<size_t>(index.col_num); j++) {
                     memcpy(old_key + offset, old_rec->data + index.cols[j].offset, index.cols[j].len);
                     offset += index.cols[j].len;
                 }
@@ -84,7 +84,7 @@ class UpdateExecutor : public AbstractExecutor {
                               .get();
                 char *new_key = new char[index.col_tot_len];
                 int offset = 0;
-                for (size_t j = 0; j < index.col_num; j++) {
+                for (size_t j = 0; j < static_cast<size_t>(index.col_num); j++) {
                     memcpy(new_key + offset, new_rec->data + index.cols[j].offset, index.cols[j].len);
                     offset += index.cols[j].len;
                 }
